Phase 3c.3 (#100): extract mapping_queries.py from mapping_service
Mechanical move of all read-only query helpers + the polyline-gap
detection helpers + the playable-trips cache out of the
~3460-LOC mapping_service.py god object into a new
services/mapping_queries.py module (~1216 LOC). Combined with
Phase 3c.1 (indexing_queue_service) and Phase 3c.2
(mapping_migrations), mapping_service.py is now down to ~2338 LOC
(was 4541 before the split started) and contains only the indexer
core: canonical_key, _index_video, index_single_file,
purge_deleted_videos, trip merge, event detection, daily stale
scan, boot catch-up, and the indexer status surface.

What moved to services.mapping_queries:
  - get_db_connection (thin wrapper around _init_db)
  - query_trips, query_trip_route, query_events, query_days,
    query_day_routes
  - playable_trips_for_date + _resolve_video_path_on_disk +
    _PLAYABLE_TRIPS_CACHE / _PLAYABLE_TRIPS_CACHE_LOCK /
    _PLAYABLE_TRIPS_TTL_SECONDS + _reset_playable_trips_cache_for_tests
  - query_all_routes_simplified
  - get_stats, get_driving_stats, get_event_chart_data
  - Polyline gap detection helpers: _haversine_m, _parse_iso_seconds,
    _is_gap_between, _simplify_polyline_rdp + GAP_MAX_*_DEFAULT

What stayed in services.mapping_service:
  - _haversine_km (used by _index_video AND clock_skew_repair)
  - _get_worker_status_for_stats and get_indexer_status (worker
    status surface, not a query)
  - All indexer/parser/canonicalize/stale-scan/boot-catchup code

Re-export contract: clean break (Phase 3c.1 pattern), NOT
re-exported from mapping_service. There were only ~13 production
caller sites (12 in blueprints/mapping.py, 1 in
cloud_archive_service.py) and ~15 test caller sites — all updated
in this PR. New code MUST import from services.mapping_queries
directly.

Dependency direction (one-way, no cycle):
  mapping_queries imports _init_db from mapping_migrations and
  _haversine_km + _get_worker_status_for_stats + _with_db_retry
  from mapping_service. mapping_service does NOT import anything
  from mapping_queries.

Testing: 1049 pass + 3 skipped (same baseline as Phase 3c.2).
Module surface verified: all public + internal helpers cleanly
moved (mapping_service no longer exposes the moved symbols).
Functional smoke: get_db_connection, query_trips, query_days,
get_stats, get_driving_stats, get_event_chart_data,
playable_trips_for_date, query_all_routes_simplified all return
sane shapes against an empty DB; gap helpers return correct
truth values.

Refs: #100 (Phase 3c.3 — read-only query helpers extracted).
Next: 3c.4 (post-deploy verify on Pi, then close #100).
