+++ .github/copilot-instructions.md
@@ -176,7 +176,7 @@ TeslaUSB has **two separate cloud upload subsystems** that share one rclone prov
 
 ## Video Indexing
 - **Single SQLite-backed queue + one worker.** All video indexing flows through `indexing_queue` (in `geodata.db`, schema v6). One low-priority background thread (`indexing_worker.py`, started inside `gadget_web.service`) drains the queue one file at a time. **Never re-introduce parallel triggers** — the old design had 6 redundant paths (every page load, every mode switch, every WiFi connect, etc.) that caused the constantly-flashing "Indexing…" banner.
-- **Module split (Phase 3c.1 + 3c.2):** the queue API (`enqueue_for_indexing`, `enqueue_many_for_indexing`, `claim_next_queue_item`, `complete_queue_item`, `defer_queue_item`, `release_claim`, `recover_stale_claims`, `compute_backoff`, `get_queue_status`, `clear_pending_queue`, `clear_all_queue`, `priority_for_path`, `_open_queue_conn`, plus `_PRIORITY_*` / `_PARSE_ERROR_*` / `_STALE_CLAIM_SECONDS` constants) lives in `services.indexing_queue_service`. The schema DDL, version constants (`_SCHEMA_VERSION`, `_BACKUP_RETENTION`, `_SCHEMA_SQL`), backup helper (`_backup_db`), `_init_db` connection factory, and the v2/v3/v4 migrations live in `services.mapping_migrations` — these are re-exported from `services.mapping_service` for backward compatibility (one-way dependency, no cycle). The cohesive indexing core (`canonical_key`, `_index_video`, `index_single_file`, `purge_deleted_videos`, trip merge, event detection, daily stale scan, boot catch-up, and the Flask query helpers — for now) stays in `services.mapping_service`. New code touching the queue MUST import from `services.indexing_queue_service`; new code touching the schema/migrations SHOULD import from `services.mapping_migrations` directly (the re-exports in `mapping_service` are for backward compat only).
+- **Module split (Phase 3c.1 + 3c.2 + 3c.3):** the queue API (`enqueue_for_indexing`, `enqueue_many_for_indexing`, `claim_next_queue_item`, `complete_queue_item`, `defer_queue_item`, `release_claim`, `recover_stale_claims`, `compute_backoff`, `get_queue_status`, `clear_pending_queue`, `clear_all_queue`, `priority_for_path`, `_open_queue_conn`, plus `_PRIORITY_*` / `_PARSE_ERROR_*` / `_STALE_CLAIM_SECONDS` constants) lives in `services.indexing_queue_service`. The schema DDL, version constants (`_SCHEMA_VERSION`, `_BACKUP_RETENTION`, `_SCHEMA_SQL`), backup helper (`_backup_db`), `_init_db` connection factory, and the v2/v3/v4 migrations live in `services.mapping_migrations` — these are re-exported from `services.mapping_service` for backward compatibility (one-way dependency, no cycle). Read-only query helpers (`get_db_connection`, `query_trips`, `query_trip_route`, `query_events`, `query_days`, `query_day_routes`, `playable_trips_for_date`, `query_all_routes_simplified`, `get_stats`, `get_driving_stats`, `get_event_chart_data`), polyline gap detection (`_haversine_m`, `_parse_iso_seconds`, `_is_gap_between`, `GAP_MAX_*_DEFAULT`, `_simplify_polyline_rdp`), the `_resolve_video_path_on_disk` filesystem probe, and the 60-second `_PLAYABLE_TRIPS_CACHE` live in `services.mapping_queries` — clean break, callers import directly. The cohesive indexing core (`canonical_key`, `_index_video`, `index_single_file`, `purge_deleted_videos`, trip merge, event detection, daily stale scan, boot catch-up, `_haversine_km`, `_get_worker_status_for_stats`, `get_indexer_status`) stays in `services.mapping_service`. New code touching the queue MUST import from `services.indexing_queue_service`; new code touching the schema/migrations SHOULD import from `services.mapping_migrations` directly (the re-exports in `mapping_service` are for backward compat only); new code touching read-only queries MUST import from `services.mapping_queries`.
 - **Producers** (the only legal ways to add work to the queue):
   - **Boot catch-up scan** — `mapping_service.boot_catchup_scan()` runs once at `gadget_web` start; cheap directory walk + batch INSERT, no parsing.
   - **inotify file watcher** — `file_watcher_service.py` callback calls `indexing_queue_service.enqueue_many_for_indexing()` on `IN_CREATE` / `IN_MOVED_TO`.
